requirements.txt
apache-beam==2.57.0
beam-mysql-connector==1.8.5